Add policy for thread deletion

+++ app/Http/Controllers/ThreadsController.php
@@ -89,6 +89,11 @@ class ThreadsController extends Controller
 
     public function destroy($channel, Thread $thread) 
     {
+        $this->authorize('update', $thread);
+        
+        if($thread->user_id != auth()->id()) {
+            abort(403, 'You do not have permission to do this.');
+        }
         $thread->delete();
 
         if(request()->wantsJson()) {
